backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
python-dotenv==0.21.0
web3==5.31.3
ipfshttpclient==0.8.0a2
scikit-learn==1.1.3
werkzeug==2.2.2